+++ engines/dfc_provider/lib/dfc_provider.rb
@@ -11,8 +11,8 @@ require "dfc_provider/supplied_product"
 require "dfc_provider/address"
 require "dfc_provider/coordination"
 
-# 🙈
-require "inspect"
+# 🙈 Monkey-patch a better inspector for semantic objects
+require "semantic_object_inspect"
 
 module DfcProvider
   DataFoodConsortium::Connector::Importer.register_type(SuppliedProduct)
