+++ engines/dfc_provider/lib/semantic_object_inspect.rb
